services/producer_simulator/requirements.txt
faker==18.9.0
aiokafka==1.4.0
pyyaml
requests
